+++ examples/access-api-with-pat.commonjs.js
@@ -1,4 +1,4 @@
-const { createAccessTokenInterceptor, createAuthClient } = require('@zitadel/node/grpc');
+const { createAccessTokenInterceptor, createAuthClient } = require('@zitadel/node/api');
 
 const apiEndpoint = 'https://zitadel-libraries-l8boqa.zitadel.cloud';
 const personalAccessToken = 'Dp6uT84aVTqYLSMpk-H1-B3SPuqFms-Tz77cgWKlyrY3JxsZcDmb7kZoRK2ixaVNfuqmXfE';
