+++ examples/access-api-with-pat.esm.mjs
@@ -1,4 +1,4 @@
-import { createAccessTokenInterceptor, createAuthClient } from '@zitadel/node/grpc';
+import { createAccessTokenInterceptor, createAuthClient } from '@zitadel/node/api';
 
 const apiEndpoint = 'https://zitadel-libraries-l8boqa.zitadel.cloud';
 const personalAccessToken = 'Dp6uT84aVTqYLSMpk-H1-B3SPuqFms-Tz77cgWKlyrY3JxsZcDmb7kZoRK2ixaVNfuqmXfE';
