+++ test/interceptors.test.ts
@@ -1,8 +1,8 @@
 import test from 'ava';
 import { CallOptions, Metadata } from 'nice-grpc';
+import { createAuthClient } from '../src/api/clients.js';
+import { createAccessTokenInterceptor, createServiceAccountInterceptor } from '../src/api/interceptors.js';
 import { ServiceAccount } from '../src/credentials/service-account.js';
-import { createAuthClient } from '../src/grpc/clients.js';
-import { createAccessTokenInterceptor, createServiceAccountInterceptor } from '../src/grpc/interceptors.js';
 import { apiEndpoint, personalAccessToken, serviceAccountJson } from './test-data.js';
 
 test('access token interceptor - create interceptor', (t) => {
